requirements.txt
beautifulsoup4==4.11.1
hydralit_components==1.0.10
joblib==1.1.0
matplotlib==3.5.2
nltk==3.7
numerize==0.12
numpy==1.21.5
pandas==1.4.4
Pillow==9.2.0
Pillow==9.5.0
plotly==5.9.0
Requests==2.31.0
scikit_learn==1.0.2
spacy==3.5.3
statsmodels==0.13.2
streamlit_lottie==0.0.5
